backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
Pillow>=10.0,<12.0
opencv-python>=4.9,<5.0
numpy>=1.26,<3.0
scikit-image>=0.22,<0.25
rembg>=2.0,<3.0
onnxruntime>=1.16,<2.0
